feat: 구조 변경 및 ts파일 추가
