--- routes/index.js
@@ -1,9 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-/* GET home page. */
-router.get('/', function (req, res, next) {
-  res.render('layout', { title: 'Express' });
-});
-
-module.exports = router;
