+++ src/routes/index.ts
@@ -1,0 +1,9 @@
+import express from "express";
+const router = express.Router();
+
+/* GET home page. */
+router.get("/", function (req: express.Request, res: express.Response, next: express.NextFunction) {
+    res.render("index", { title: "Express" });
+});
+
+export default router;
