+++ src/routes/users.ts
@@ -1,0 +1,9 @@
+import express from "express";
+const router = express.Router();
+
+/* GET users listing. */
+router.get("/", function (req: express.Request, res: express.Response, next: express.NextFunction) {
+    res.send("respond with a resource");
+});
+
+export default router;
